fixes in different between TestsSourceSnapshot and TestsSourceVersion

+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/controllers/OrganizationController.kt
@@ -512,14 +512,27 @@ internal class OrganizationController(
         }
 
     private fun cleanupStorageData(testSuite: TestSuite) = blockingToMono {
-        testsSourceVersionService.findSnapshot(
+        testsSourceVersionService.delete(
             testSuite.source.organization.name,
             testSuite.source.name,
             testSuite.version,
         )
     }
-        .flatMap { testsSourceSnapshotStorage.delete(it) }
-        .defaultIfEmpty(false)
+        .flatMap { noVersionForSnapshot ->
+            if (noVersionForSnapshot) {
+                blockingToMono {
+                    testsSourceVersionService.findSnapshot(
+                        testSuite.source.organization.name,
+                        testSuite.source.name,
+                        testSuite.version,
+                    )
+                }
+                    .flatMap { testsSourceSnapshotStorage.delete(it) }
+                    .defaultIfEmpty(false)
+            } else {
+                true.toMono()
+            }
+        }
 
     private fun getFilteredOrganizationDtoList(filters: OrganizationFilters): Flux<OrganizationDto> = blockingToFlux {
         organizationService.getFiltered(filters)
